work in progress on implementing the marshalling #318
Fixing the compilation.

Finished with marshalling of all supported curves #318

remove comment

rm curve25519
