+++ libs/algebra/include/nil/crypto3/algebra/curves/detail/babyjubjub/g1.hpp
@@ -31,7 +31,6 @@
 
 #include <nil/crypto3/algebra/curves/forms.hpp>
 #include <nil/crypto3/algebra/curves/detail/forms/twisted_edwards/element_g1_affine.hpp>
-#include <nil/crypto3/algebra/curves/detail/forms/montgomery/element_g1_affine.hpp>
 
 namespace nil {
     namespace crypto3 {
